[Video 83] Context API With Custom Hook (Avoid Prop Drilling Pt 2)
