+++ trekbag/src/components/Header.jsx
@@ -1,11 +1,16 @@
 import Counter from "./Counter";
 import Logo from "./Logo";
+import { useItemsContext } from "./lib/hooks";
 
-export default function Header({ totalItems, itemsPacked }) {
+export default function Header() {
+  const { items } = useItemsContext();
   return (
     <header>
       <Logo />
-      <Counter totalItems={totalItems} itemsPacked={itemsPacked} />
+      <Counter
+        totalItems={items.length}
+        itemsPacked={items.filter((item) => item.checked).length}
+      />
     </header>
   );
 }
