+++ trekbag/src/components/lib/hooks.js
@@ -1,0 +1,13 @@
+import { useContext } from "react";
+import { ItemsContext } from "../../contexts/ItemsContextProvider";
+
+export function useItemsContext() {
+  const context = useContext(ItemsContext);
+
+  if (!context) {
+    throw new Error(
+      "useItemsContext must be used within an ItemsContextProvider"
+    );
+  }
+  return context;
+}
